categorize malfind as malware plugin
